Share one error-block formatter for AssertionScope and ShouldSatisfyAllConditions
Both rendered the numbered "Error N" blocks independently and had drifted:
ShouldSatisfyAllConditions used 39-wide error headers against a 41-wide
closing divider, while AssertionScope put the blank line before each error
rather than after. Extract a single AssertionErrorFormatter that both call,
so they share the fixed-width centred dividers and the blank-line-after-each-
error layout and cannot drift apart again.
