+++ src/DocumentationExamples/CodeExamples/AssertionScopeExamples.BasicUsage.exceptionText.approved.txt
@@ -16,5 +16,6 @@
     1000000
         but was
     30000
+
 -----------------------------------------
 ```
