+++ src/DocumentationExamples/CodeExamples/ShouldSatisfyAllConditionsExamples.ShouldSatisfyAllConditions.exceptionText.approved.txt
@@ -2,11 +2,11 @@
 mrBurns
     should satisfy all the conditions specified, but does not.
 The following errors were found ...
---------------- Error 1 ---------------
+---------------- Error 1 ----------------
     mrBurns.Name (null)
         should not be null or empty
 
---------------- Error 2 ---------------
+---------------- Error 2 ----------------
     mrBurns.Name
         should be
     "Mr.Burns"
