+++ src/Shouldly.Tests/AssertionScopes/AssertionScopeScenarios.MixedPassAndFail_OnlyCollectsFailures.approved.txt
@@ -13,4 +13,5 @@
         difference
     Expected: "b"
     Actual:   "a"
+
 -----------------------------------------
