+++ src/Shouldly.Tests/AssertionScopes/AssertionScopeScenarios.MultipleFailures_ShouldCollectAll_WithoutEnclosingBraces.approved.txt
@@ -13,4 +13,5 @@
         difference
     Expected: "expected"
     Actual:   "actual"
+
 -----------------------------------------
