+++ src/Shouldly.Tests/AssertionScopes/AssertionScopeScenarios.NestedScope_PropagatesFailuresToOuterScope.approved.txt
@@ -13,4 +13,5 @@
         difference
     Expected: "outer"
     Actual:   "inner"
+
 -----------------------------------------
