+++ src/Shouldly.Tests/AssertionScopes/AssertionScopeScenarios.ShouldSatisfyAllConditions_InsideScope_WorksCorrectly.approved.txt
@@ -4,4 +4,5 @@
         should be
     2
         but was not
+
 -----------------------------------------
