+++ src/Shouldly.Tests/AssertionScopes/AssertionScopeScenarios.SingleFailure_ShouldThrowWithMessage.approved.txt
@@ -4,4 +4,5 @@
         should be
     2
         but was not
+
 -----------------------------------------
