+++ src/Shouldly/AssertionScope.cs
@@ -57,42 +57,10 @@ public sealed class AssertionScope : IDisposable
         }
     }
 
-    private const int DividerWidth = 41;
-
     private string BuildMessage()
     {
-        var sb = new StringBuilder();
-        var errorCount = 1;
         var failureCount = failures.Count;
-        sb.AppendLine($"{failureCount} assertion{(failureCount == 1 ? "" : "s")} in this scope failed:");
-        foreach (var failure in failures)
-        {
-            if (errorCount > 1)
-                sb.AppendLine();
-            sb.AppendLine(Divider($"Error {errorCount}"));
-            var lines = failure.Message?.Replace("\r\n", "\n").Split('\n') ?? [];
-            var paddedLines = lines.Select(l => string.IsNullOrEmpty(l) ? l : "    " + l);
-            sb.AppendLine(string.Join(Environment.NewLine, paddedLines.ToArray()));
-            errorCount++;
-        }
-
-        sb.AppendLine(Divider());
-
-        return sb.ToString().TrimEnd();
-    }
-
-    /// <summary>
-    /// Builds a divider line of a fixed width so the error headers and the closing
-    /// divider always align in monospaced output, regardless of the error number.
-    /// </summary>
-    private static string Divider(string? label = null)
-    {
-        if (string.IsNullOrEmpty(label))
-            return new string('-', DividerWidth);
-
-        var text = $" {label} ";
-        var dashes = Math.Max(DividerWidth - text.Length, 2);
-        var left = dashes / 2;
-        return new string('-', left) + text + new string('-', dashes - left);
+        var header = $"{failureCount} assertion{(failureCount == 1 ? "" : "s")} in this scope failed:";
+        return header + Environment.NewLine + AssertionErrorFormatter.FormatErrors(failures.Select(f => f.Message));
     }
 }
